Add support for the new Internship class
Added the new Internship class, along with InternshipDescription, InternshipName and InternshipInterviewDateTime classes to serve as objects representing attributes of the Internship class.

Modified the JsonAdaptedCompany class and added the JsonAdaptedInternship class to support loading of Internship data from addressbook.json.

Modified SampleDataUtil to add in the internship data for a sample addressbook.

Modified the Company constructor and variables, as well as existing company related commands (EditCompanyCommand, AddCommandParser) to accommodate the new Internship data.

Added support for formatting Companies and Internships by modifying the formatCompany command and introducing the formatInternship command in Messages.java.

Added the DateTimeParserUtil class to parse any given String or LocalDateTime object, provided that the given object is valid.